+++ CameraServer/Models/RequestBodies/MoveCameraBody.cs
@@ -5,8 +5,8 @@ namespace CameraServer.Models.RequestBodies
     public class MoveCameraBody : RequestBody
     {
         public int CameraId { get; set; }
-        public int DeltaPitch { get; set; }
-        public int DeltaYaw { get; set; }
+        public float DeltaPitch { get; set; }
+        public float DeltaYaw { get; set; }
 
         public override bool Valid => CameraId > 0;
     }
